REFACTOR::
Now I think is more OOP, the SuperTextMenu is my final version, can be used as a library, and in this case implements Runnable
that make more easy create submenus, clear, easy and fast to build. to make more insolated don't use the view to print the menu
making it less MVC but wining in aislation. can be used in other proyects easy, or refactor to use a view class.